.skills/zeyu-docs-retrieval/requirements.txt
requests>=2.31.0
pyyaml>=6.0
numpy>=1.24.0
sentence-transformers>=2.2.0
